formatted both frontend and backend

+++ frontend/src/services/addressService.js
@@ -15,15 +15,11 @@ export const getAddresses = async () => {
 export const addAddress = async (address) => {
   const token = localStorage.getItem("access");
 
-  const response = await API.post(
-    "/auth/addresses/",
-    address,
-    {
-      headers: {
-        Authorization: `Bearer ${token}`,
-      },
-    }
-  );
+  const response = await API.post("/auth/addresses/", address, {
+    headers: {
+      Authorization: `Bearer ${token}`,
+    },
+  });
 
   return response.data;
 };
@@ -31,15 +27,11 @@ export const addAddress = async (address) => {
 export const updateAddress = async (id, address) => {
   const token = localStorage.getItem("access");
 
-  const response = await API.put(
-    `/auth/addresses/${id}/`,
-    address,
-    {
-      headers: {
-        Authorization: `Bearer ${token}`,
-      },
-    }
-  );
+  const response = await API.put(`/auth/addresses/${id}/`, address, {
+    headers: {
+      Authorization: `Bearer ${token}`,
+    },
+  });
 
   return response.data;
 };
@@ -47,12 +39,9 @@ export const updateAddress = async (id, address) => {
 export const deleteAddress = async (id) => {
   const token = localStorage.getItem("access");
 
-  await API.delete(
-    `/auth/addresses/${id}/`,
-    {
-      headers: {
-        Authorization: `Bearer ${token}`,
-      },
-    }
-  );
+  await API.delete(`/auth/addresses/${id}/`, {
+    headers: {
+      Authorization: `Bearer ${token}`,
+    },
+  });
 };
